+++ models/linkedin_ads__campaign_report.sql
@@ -65,7 +65,7 @@ final as (
         campaign.run_schedule_end_at,
         campaign.last_modified_at,
         campaign.created_at,
-        report.total_conversions,
+        sum(report.total_conversions) as total_conversions,
         sum(report.clicks) as clicks,
         sum(report.impressions) as impressions,
         sum(report.cost) as cost,
@@ -86,7 +86,7 @@ final as (
         on campaign.account_id = account.account_id
         and campaign.source_relation = account.source_relation
 
-    {{ dbt_utils.group_by(31) }}
+    {{ dbt_utils.group_by(30) }}
 
 )